refactor(FakerError): move into errors directory
