+++ src/faker.ts
@@ -7,13 +7,13 @@ import { Datatype } from './datatype';
 import { _Date } from './date';
 import type { LocaleDefinition } from './definitions';
 import { DEFINITIONS } from './definitions';
+import { FakerError } from './errors/faker.error';
 import { Fake } from './fake';
 import { Finance } from './finance';
 import { Git } from './git';
 import { Hacker } from './hacker';
 import { Helpers } from './helpers';
 import { Image } from './image';
-import { FakerError } from './internal/faker.error';
 import { Internet } from './internet';
 import type { KnownLocale } from './locales';
 import { Lorem } from './lorem';
