+++ src/mersenne.ts
@@ -1,4 +1,4 @@
-import { FakerError } from './internal/faker.error';
+import { FakerError } from './errors/faker.error';
 import Gen from './utils/mersenne';
 
 /**
