+++ test/finance.spec.ts
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, it } from 'vitest';
 import { faker } from '../src';
+import { FakerError } from '../src/errors/faker.error';
 import ibanLib from '../src/iban';
-import { FakerError } from '../src/internal/faker.error';
 import { luhnCheck } from './support/luhnCheck';
 
 const seedRuns = [
